requirements.txt
beautifulsoup4==4.11.1
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==1.4.39
